chore(kumactl) remove kumactl plugin hooks
The only user of the kumactl plugin hooks was the gateway plugin.
Improved protobuf codegen means that we don't need this mechanism to
register gateway types for kumactl any more, and we can consolidate the
registration code in the gateway plugin.

This fixes #2503.
