+++ pkg/plugins/runtime/gateway/register/doc.go
@@ -1,0 +1,1 @@
+package register
